+++ core-tests/e2e-tests/spring/spring-rest-openapi-v3/src/test/kotlin/org/evomaster/e2etests/spring/openapi/v3/SpringTestBase.kt
@@ -3,8 +3,16 @@ package org.evomaster.e2etests.spring.openapi.v3
 import org.evomaster.client.java.controller.InstrumentedSutStarter
 import org.evomaster.client.java.instrumentation.InputProperties
 import org.evomaster.client.java.instrumentation.InstrumentingAgent
+import org.evomaster.client.java.instrumentation.shared.ClassName
 import org.evomaster.e2etests.utils.RestTestBase
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeAll
+import java.util.Optional.ofNullable
+import kotlin.reflect.KClass
+import kotlin.reflect.KMutableProperty1
+import kotlin.reflect.full.createInstance
+import kotlin.reflect.full.memberProperties
+import kotlin.reflect.jvm.isAccessible
 
 /**
  * Created by arcuri82 on 03-Mar-20.
@@ -23,6 +31,25 @@ abstract class SpringTestBase : RestTestBase(){
             InstrumentedSutStarter.loadAgent()
             InstrumentingAgent.changePackagesToInstrument("com.foo.")
         }
+    }
+
+    protected fun initDtoClass(name: String): Pair<KClass<out Any>, Any> {
+        val className = ClassName("org.foo.dto.$name")
+        val klass = loadClass(className).kotlin
+        Assertions.assertNotNull(klass)
+        return Pair(klass, klass.createInstance())
+    }
+
+    protected fun assertProperty(klass: KClass<out Any>, instance: Any, propertyName: String, propertyValue: Any?) {
+        val property = klass.memberProperties
+            .firstOrNull { it.name == propertyName } as? KMutableProperty1<Any, Any?>
+        Assertions.assertNotNull(property)
+
+        property?.let {
+            it.isAccessible = true
+            it.set(instance, ofNullable(propertyValue))
+        }
+        Assertions.assertEquals(ofNullable(propertyValue), property?.get(instance))
     }
 
 }
