+++ core-tests/e2e-tests/spring/spring-rest-openapi-v3/src/test/kotlin/org/evomaster/e2etests/spring/openapi/v3/jsonpatchstringvalue/JsonPatchStringValueDtoEMTest.kt
@@ -1,18 +1,11 @@
 package org.evomaster.e2etests.spring.openapi.v3.jsonpatchstringvalue
 
 import com.foo.rest.examples.spring.openapi.v3.jsonpatchstringvalue.JsonPatchStringValueController
-import org.evomaster.client.java.instrumentation.shared.ClassName
 import org.evomaster.core.problem.rest.data.HttpVerb
 import org.evomaster.e2etests.spring.openapi.v3.SpringTestBase
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
-import java.util.Optional.ofNullable
-import kotlin.reflect.KClass
-import kotlin.reflect.KMutableProperty1
-import kotlin.reflect.full.createInstance
-import kotlin.reflect.full.memberProperties
-import kotlin.reflect.jvm.isAccessible
 
 /**
  * E2E test: JSON Patch DTO with a PRIMITIVE (string) value.
@@ -50,6 +43,10 @@ class JsonPatchStringValueDtoEMTest : SpringTestBase() {
         ) { args: MutableList<String> ->
 
             setOption(args, "dtoForRequestPayload", "true")
+            // This SUT intentionally has no authentication, and the PATCH returns 2xx, so the
+            // security oracle would flag a fault 208 (Anonymous Modifications). It is irrelevant to
+            // what we assert here (JSON Patch DTO generation), so we disable it to keep the suite clean.
+            setOption(args, "security", "false")
 
             val solution = initAndRun(args)
 
@@ -71,24 +68,5 @@ class JsonPatchStringValueDtoEMTest : SpringTestBase() {
         assertProperty(klass, instance, "from", "/description")
         // A JSON Patch value can be any JSON value; here it is a primitive string inlined as a literal.
         assertProperty(klass, instance, "value", "EvoMaster")
-    }
-
-    private fun initDtoClass(name: String): Pair<KClass<out Any>, Any> {
-        val className = ClassName("org.foo.dto.$name")
-        val klass = loadClass(className).kotlin
-        Assertions.assertNotNull(klass)
-        return Pair(klass, klass.createInstance())
-    }
-
-    private fun assertProperty(klass: KClass<out Any>, instance: Any, propertyName: String, propertyValue: Any?) {
-        val property = klass.memberProperties
-            .firstOrNull { it.name == propertyName } as? KMutableProperty1<Any, Any?>
-        Assertions.assertNotNull(property)
-
-        property?.let {
-            it.isAccessible = true
-            it.set(instance, ofNullable(propertyValue))
-        }
-        Assertions.assertEquals(ofNullable(propertyValue), property?.get(instance))
     }
 }
